pyproject.toml
[tool.poetry]
name = "user-part"
version = "0.1.0"
description = ""
authors = ["Sergey <ssnovak@stud.kpfu.ru>"]
readme = "README.md"
packages = [{include = "user_part"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
psycopg = "^3.1.9"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
